Move setup of toolbox to internal header.
Done to make it easy to construct multiple example/test programs.
